refactor: add kakaoApi interface
