+++ src/main/java/gift/service/OrderService.java
@@ -26,31 +26,20 @@ import java.util.Optional;
 
 @Service
 public class OrderService {
-    private static final Logger logger = LoggerFactory.getLogger(KakaoAuthService.class);
-
-    private final RestClient client;
-
+    private final KakaoApi kakaoApi;
     private final OptionService optionService;
     private final ProductService productService;
     private final KakaoAuthService kakaoAuthService;
     private final PointService pointService;
     private final WishRepository wishRepository;
 
-    public OrderService(RestClient.Builder builder, KakaoProperties kakaoProperties, OptionService optionService, ProductService productService, KakaoAuthService kakaoAuthService, PointService pointService, WishRepository wishRepository) {
+    public OrderService(KakaoApi kakaoApi, OptionService optionService, ProductService productService, KakaoAuthService kakaoAuthService, PointService pointService, WishRepository wishRepository) {
+        this.kakaoApi = kakaoApi;
         this.optionService = optionService;
         this.productService = productService;
         this.kakaoAuthService = kakaoAuthService;
         this.pointService = pointService;
         this.wishRepository = wishRepository;
-
-        SimpleClientHttpRequestFactory requestFactory = new SimpleClientHttpRequestFactory();
-        requestFactory.setConnectTimeout(kakaoProperties.getConnectTimeout());
-        requestFactory.setReadTimeout(kakaoProperties.getReadTimeout());
-
-        this.client = builder
-                .requestFactory(requestFactory)
-                .requestInterceptor(new ClientInterceptor())
-                .build();
     }
 
     public void orderProduct(String token, OrderDTO orderDTO) throws JsonProcessingException {
@@ -72,42 +61,6 @@ public class OrderService {
         }
 
         //카카오톡 메세지 보내기
-        sendKakaoMessage(token, orderDTO.getMessage());
-    }
-
-    public void sendKakaoMessage(String token, String message) throws JsonProcessingException {
-        logger.info("sendKakaoMessage");
-        var url = "https://kapi.kakao.com/v2/api/talk/memo/default/send";
-        var headers = new HttpHeaders();
-        headers.setBearerAuth(token);
-        headers.setContentType(MediaType.APPLICATION_FORM_URLENCODED);
-
-        var body = createTemplateObject(message);
-
-        var response = this.client.post()
-                .uri(url)
-                .headers(httpHeaders -> {
-                    httpHeaders.addAll(headers);
-                })
-                .body(body)
-                .retrieve()
-                .toEntity(String.class);
-
-        logger.info("sendKakaoMessget result" + response);
-    }
-    private @NotNull LinkedMultiValueMap<String, String> createTemplateObject(String message) throws JsonProcessingException {
-        String objectType = "text";
-        String webUrl = "http://localhost:8080/";
-        String buttonTitle = "바로가기";
-
-        LinkObject link = new LinkObject(webUrl);
-        TemplateObject templateObject = new TemplateObject(objectType, message, link, buttonTitle);
-
-        String jsonString = new ObjectMapper().writeValueAsString(templateObject);
-
-        LinkedMultiValueMap<String, String> body = new LinkedMultiValueMap<>();
-        body.add("template_object", jsonString);
-
-        return body;
+        kakaoApi.sendKakaoMessage(token, orderDTO.getMessage());
     }
 }
